API/requirements.txt
aiofiles==22.1.0
anyio==3.6.1
autopep8==1.7.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
fastapi==0.85.0
google-api-core==2.10.1
google-auth==2.12.0
google-cloud==0.34.0
google-cloud-vision==3.1.3
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-status==1.49.1
h11==0.14.0
httptools==0.5.0
idna==3.4
numpy==1.23.3
opencv-python==4.6.0.66
proto-plus==1.22.1
protobuf==4.21.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pydantic==1.10.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
toml==0.10.2
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3